Refactor: 체험 수정 api 컴포넌트 분리 작업
